Code clean-up: int to size_t

+++ src/tokenizer/Scanner.cpp
@@ -13,7 +13,7 @@ std::vector<Token> Scanner::scanTokens(){
 }
 
 bool Scanner::isAtEnd(){
-  return current >= static_cast<int>(source.length());
+  return current >= source.length();
 }
 
 void Scanner::addToken(TokenType type){
@@ -21,13 +21,13 @@ void Scanner::addToken(TokenType type){
 }
 
 void Scanner::addToken(TokenType type, std::any literal){
-  std::string text{source.substr(static_cast<size_t>(start), static_cast<size_t>(current - start))};
+  std::string text{source.substr(start, current - start)};
   tokens.emplace_back(type, text, literal, line);
 }
 
 void Scanner::identifier(){
   while(isAlphaNumeric(peek())) advance();
-  std::string text{source.substr(static_cast<size_t>(start), static_cast<size_t>(current - start))};
+  std::string text{source.substr(start, current - start)};
   auto it = keywords.find(text);
   TokenType type = it == keywords.end() ? TokenType::IDENTIFIER : it->second;
   addToken(type);
@@ -41,7 +41,7 @@ void Scanner::number(){
     while(isDigit(peek())) advance();
   }
 
-  std::string text{source.substr(static_cast<size_t>(start), static_cast<size_t>(current - start))};
+  std::string text{source.substr(start, current - start)};
   double number = std::stod(std::string{text});
   addToken(TokenType::NUMBER, number);
 }
@@ -64,19 +64,19 @@ void Scanner::string(){
 }
 
 bool Scanner::match(char expected){
-  if (isAtEnd() || source.at(static_cast<size_t>(current)) != expected) return false;
+  if (isAtEnd() || source.at(current) != expected) return false;
   current++;
   return true;
 }
 
 char Scanner::peek(){
   if(isAtEnd()) return '\0';
-  return source.at(static_cast<size_t>(current));
+  return source.at(current);
 }
 
 char Scanner::peekNext(){
-  if(current + 1 > static_cast<int>(source.length())) return '\0';
-  return source.at(static_cast<size_t>(current + 1));
+  if(current + 1 > source.length()) return '\0';
+  return source.at(current + 1);
 }
 
 bool Scanner::isAlpha(char c){
@@ -94,7 +94,7 @@ bool Scanner::isDigit(char c){
 }
 
 char Scanner::advance(){
-  return source[static_cast<size_t>(current++)];
+  return source[current++];
 }
 
 void Scanner::scanToken(){
